+++ doc/changelog.rst
@@ -2,6 +2,8 @@
 Changelog
 =========
 
+- :bug:`344` :func:`galgebra.metric.collect` no longer discards terms that were not requested.
+
 - :support:`338` The undocumented and mispelt static method :meth:`galgebra.lt.Mlt.extact_basis_indexes` (which just computed a value equivalent to :attr:`galgebra.ga.Ga.basis_super_scripts`) has been deprecated.
 
 - :feature:`326` The :meth:`galgebra.ga.Ga.make_grad` function now accepts multivectors, not just vectors.